% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Leveraging Implementation}
\usage{
algo_leverage(y, X, sample_size, method = "uniform")
}
\arguments{
\item{y}{The response in vector form}

\item{X}{The covariate in matrix form}

\item{sample_size}{sample size for leveraging}

\item{method}{The method to implement, 'weighted' or 'uniform'}
}
\value{
coefficients of the linear regression model
}
\description{
Implements of algorithmic leveraging for linear regression using uniform and leverage score based subsampling of rows
}
