% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{Linear system solver}
\usage{
solve_ols(
  A,
  b,
  x0,
  num_cores = 1,
  parallel = F,
  method = "Gauss",
  maxit = 10000
)
}
\arguments{
\item{A}{The A part for linear system of Ax = b, a matrix object}

\item{b}{The b part for linear system of Ax = b, a matrix object}

\item{x0}{Initial value of the solution}

\item{num_cores}{Number of cores for parallel Jacobi.}

\item{parallel}{TRUE for applying parallel computing for Jacobi method}

\item{method}{The method used for solver: "Gauss" or "Jacobi"}

\item{maxit}{Maximum Number of Iteration}
}
\value{
The solution to the linear system
}
\description{
Iterative Methods for Solving Linear System of Equations
}
